+++ System.Geometry.Text/GlyphMetrics.cs
@@ -13,20 +13,20 @@ namespace System.Geometry.Text
         /// <summary>
         /// The glyph advance; this is the distance to advance the pen after positioning the glyph.
         /// </summary>
-        public readonly float Advance;
+        public readonly double Advance;
 
         /// <summary>
         /// The linear advance; this is the original advance distance unaffected by hinting.
         /// Use this if you want a guaranteed smooth transition of advances across various pixel sizes.
         /// </summary>
-        public readonly float LinearAdvance;
+        public readonly double LinearAdvance;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="GlyphMetrics"/> struct.
         /// </summary>
         /// <param name="advance">The advance distance.</param>
         /// <param name="linearAdvance">The linear unhinted advance distance.</param>
-        internal GlyphMetrics(float advance, float linearAdvance)
+        internal GlyphMetrics(double advance, double linearAdvance)
         {
             Advance = advance;
             LinearAdvance = linearAdvance;
